requirements.txt
python-dotenv~=1.1.1
requests~=2.32.4
ollama>=0.1.14 